requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-appconf==1.0.4
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
packaging==20.9
pilkit==2.0
Pillow==8.2.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.17.1
python-magic==0.4.24
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
